+++ R/PersonClass.R
@@ -1,6 +1,7 @@
 Person <- setRefClass("Person", 
                       
-                      fields = list(raw_responses = 'character',
+                      fields = list(ID = 'integer',
+                                    raw_responses = 'character',
                                     responses = 'integer',
                                     items_answered = 'integer',
                                     thetas = 'matrix',
@@ -11,12 +12,15 @@ Person <- setRefClass("Person",
                                     item_time = 'numeric',
                                     valid_item = 'logical',
                                     login_name = 'character',
-                                    score = 'logical'),
+                                    score = 'logical',
+                                    true_thetas = 'numeric'),
                       
                       methods = list(
                          initialize = function(nfact, nitems, thetas.start_in, score,
-                                               theta_SEs){
+                                               theta_SEs, ID = 0L){
                              'Initialize the person object given background information'
+                             ID <<- ID
+                             true_thetas <<- numeric(0L)
                              raw_responses <<- as.character(rep(NA, nitems))
                              responses <<- as.integer(rep(NA, nitems))
                              valid_item <<- rep(TRUE, nitems)