+++ R/mirtCAT-package.R
@@ -26,5 +26,7 @@
 #' @useDynLib mirtCAT
 #' @import mirt shiny methods lattice Rcpp
 #' @importFrom stats approx na.omit pnorm qnorm reshape runif spline
+#' @importFrom lpSolve lp
+#' @importFrom pbapply pblapply
 #' @keywords package
 NULL